Rename enrolments folder as signups
