+++ src/domain/signups/buttonPanel/__tests__/ButtonPanel.test.tsx
