+++ src/domain/signups/query.ts
