+++ src/domain/signups/searchPanel/__tests__/SearchPanel.test.tsx
