+++ src/domain/signups/searchPanel/searchPanel.module.scss
